+++ includes/prep_comps.h
@@ -13,7 +13,7 @@
 # define PREP_COMPS_H
 
 void	ft_prepare_computations(t_comps *comps, t_intersection *closest, const t_ray *ray);
-t_vct	ft_ray_point(const t_ray *ray, float t);
+t_vct	ft_ray_point(const t_ray *ray, double t);
 t_vct	ft_reflect(const t_vct *in, const t_vct *normal);
 t_vct	ft_over_point(const t_vct *point, const t_vct *normal);
